+++ tests/test_moldrug.py
@@ -352,7 +352,14 @@ def test_constraintconf():
         fix= os.path.join(tmp_path.name, 'fix.sdf'),
         out = os.path.join(tmp_path.name, 'conf.sdf')
     )
+def test_generate_conformers():
+    from moldrug.constraintconf import generate_conformers
+    from rdkit.Chem import AllChem
+    ref = Chem.MolFromSmiles('O=S(=O)(Nc1ccc(Cl)cc1)c1ccsc1C(O)O')
+    mol = Chem.MolFromSmiles('CN(C)S(=O)(=O)c1cc(NS(=O)(=O)c2ccsc2C(O)O)ccc1Cl')
 
-
+    AllChem.EmbedMolecule(ref) 
+    AllChem.MMFFOptimizeMolecule(ref)
+    generate_conformers(Chem.RemoveHs(mol), Chem.RemoveHs(ref), 50)
 if __name__ == '__main__':
     pass